requirements.txt
openai-whisper==20231117
moviepy==1.0.3
gradio==4.38.1